OP controlled stop and go
Defualt behavior is to trigger the factory `standstill` mode which requires user input for the vehicle to resume. Preventing standstill mode will allow OP to control brake and acceleration without the vehicle going into standstill mode, with the caveat being that acceleration commands CANNOT be sent until the vehicle is moving.

This means that with this "hack" enabled, stop and go will work, user input is not required for the vehicle to begin moving again BUT if the vehicle cannot creep forward when the brake is released, OP will not be able to send acceleration commands due to the factory ECU locking accel out.

TL;DR this enables stop and go controlled by OP.