If mpc and cruise both equal choose mpc
When coming to a full stop at a traffic light and a lead car is detected.
The braking will be for the lead car and not for the traffic light i.e. where the lead car is ignored.

(arne182/openpilot/071-clean/5f5c63307ca9011c1fb82bc4db00197050a5d825)